refactor: WorkshopComponents.swift -> 거의 총기분해급으로 해체쇼
- WorkshopSortSheet.swift (SortOption 포함)
- WorkshopFilterBar.swift
- WorkshopFilterChip.swift
- WorkshopSkeletonBox.swift
- WorkshopItemCard.swift (PriceOverlay, SoundPlayButton, DownloadProgressRing 포함)
- WorkshopItemActionButton.swift
