+++ Keychy/Keychy/Presentation/Workshop/Views/Components/WorkshopSkeletonBox.swift
@@ -1,0 +1,47 @@
+//
+//  WorkshopSkeletonBox.swift
+//  Keychy
+//
+//  Created by 길지훈 on 1/22/26.
+//
+
+import SwiftUI
+
+// MARK: - Skeleton Loading View
+
+struct WorkshopSkeletonBox: View {
+    let width: CGFloat
+    let height: CGFloat
+
+    @State private var isAnimating = false
+
+    var body: some View {
+        RoundedRectangle(cornerRadius: 10)
+            .fill(Color.gray50)
+            .frame(width: width, height: height)
+            .overlay(
+                RoundedRectangle(cornerRadius: 10)
+                    .fill(
+                        LinearGradient(
+                            gradient: Gradient(colors: [
+                                Color.clear,
+                                Color.white.opacity(0.5),
+                                Color.clear
+                            ]),
+                            startPoint: .leading,
+                            endPoint: .trailing
+                        )
+                    )
+                    .offset(x: isAnimating ? width : -width)
+                    .animation(
+                        Animation.linear(duration: 1.5)
+                            .repeatForever(autoreverses: false),
+                        value: isAnimating
+                    )
+            )
+            .clipShape(RoundedRectangle(cornerRadius: 10))
+            .onAppear {
+                isAnimating = true
+            }
+    }
+}
